:truck: rename mapper files and add comments
