+++ lib/requesters/reviewsMappedRequests.js
@@ -5,27 +5,41 @@ const R = require('ramda');
 const request = require('../utils/request');
 const scriptData = require('../utils/scriptData');
 const reviewsList = require('../utils/reviewsList');
-const { REQUEST_MAPPINGS } = require('../mappers/request');
+const { REQUEST_MAPPINGS } = require('../mappers/reviews');
 const { BASE_URL } = require('../utils/configurations');
 
-const TOKEN_TYPES = {
+/**
+ * This object allow us to differ between
+ * the initial body request and the paginated ones
+ */
+const REQUEST_TYPE = {
   initial: 'initial',
-  request: 'request'
+  paginated: 'paginated'
 };
 
+/**
+ * This method allow us to get the body for the review request
+ *
+ * @param options.appId The app id for reviews
+ * @param options.sort The sort order for reviews
+ * @param options.numberOfReviewsPerRequest The number of reviews per request
+ * @param options.withToken The token to be used for the given request
+ * @param options.requestType The
+ */
 function getBodyForRequests ({
   appId,
   sort,
-  numberOfReviews = 100,
+  numberOfReviewsPerRequest = 100,
   withToken = '%token%',
-  tokenType = 'initial'
+  requestType = REQUEST_TYPE.initial
 }) {
+  /* The body is slight different for the initial and paginated requests */
   const formBody = {
-    initial: `f.req=%5B%5B%5B%22UsvDTd%22%2C%22%5Bnull%2Cnull%2C%5B2%2C${sort}%2C%5B${numberOfReviews}%2Cnull%2Cnull%5D%2Cnull%2C%5B%5D%5D%2C%5B%5C%22${appId}%5C%22%2C7%5D%5D%22%2Cnull%2C%22generic%22%5D%5D%5D`,
-    request: `f.req=%5B%5B%5B%22UsvDTd%22%2C%22%5Bnull%2Cnull%2C%5B2%2C${sort}%2C%5B${numberOfReviews}%2Cnull%2C%5C%22${withToken}%5C%22%5D%2Cnull%2C%5B%5D%5D%2C%5B%5C%22${appId}%5C%22%2C7%5D%5D%22%2Cnull%2C%22generic%22%5D%5D%5D`
+    [REQUEST_TYPE.initial]: `f.req=%5B%5B%5B%22UsvDTd%22%2C%22%5Bnull%2Cnull%2C%5B2%2C${sort}%2C%5B${numberOfReviewsPerRequest}%2Cnull%2Cnull%5D%2Cnull%2C%5B%5D%5D%2C%5B%5C%22${appId}%5C%22%2C7%5D%5D%22%2Cnull%2C%22generic%22%5D%5D%5D`,
+    [REQUEST_TYPE.paginated]: `f.req=%5B%5B%5B%22UsvDTd%22%2C%22%5Bnull%2Cnull%2C%5B2%2C${sort}%2C%5B${numberOfReviewsPerRequest}%2Cnull%2C%5C%22${withToken}%5C%22%5D%2Cnull%2C%5B%5D%5D%2C%5B%5C%22${appId}%5C%22%2C7%5D%5D%22%2Cnull%2C%22generic%22%5D%5D%5D`
   };
 
-  return formBody[tokenType];
+  return formBody[requestType];
 }
 
 async function processAndRecur (html, opts, savedReviews, mappings) {
@@ -34,8 +48,8 @@ async function processAndRecur (html, opts, savedReviews, mappings) {
   }
 
   const reviews = reviewsList.extract(mappings.reviews, html, opts.appId);
-  const token = R.path(mappings.reviewsToken, html);
-  opts.tokenType = TOKEN_TYPES.request;
+  const token = R.path(mappings.token, html);
+  opts.requestType = REQUEST_TYPE.paginated;
 
   return checkFinished(opts, [...savedReviews, ...reviews], token);
 }
@@ -43,7 +57,7 @@ async function processAndRecur (html, opts, savedReviews, mappings) {
 function checkFinished (opts, savedReviews, nextToken) {
   debug('nextToken: %s', nextToken);
   debug('savedReviews length: %s', savedReviews.length);
-  debug('tokenType: %s', opts.tokenType);
+  debug('requestType: %s', opts.requestType);
 
   if (savedReviews.length >= opts.num || !nextToken) {
     return savedReviews.slice(0, opts.num);
@@ -53,7 +67,7 @@ function checkFinished (opts, savedReviews, nextToken) {
     appId: opts.appId,
     sort: opts.sort,
     withToken: nextToken,
-    tokenType: opts.tokenType
+    requestType: opts.requestType
   });
   const url = `${BASE_URL}/_/PlayStoreUi/data/batchexecute?rpcids=qnKhOb&f.sid=-697906427155521722&bl=boq_playuiserver_20190903.08_p0&hl=${opts.lang}&gl=${opts.country}&authuser&soc-app=121&soc-platform=1&soc-device=1&_reqid=1065213`;
 
@@ -82,7 +96,7 @@ function checkFinished (opts, savedReviews, nextToken) {
 }
 
 function processFullReviews (opts) {
-  opts.tokenType = TOKEN_TYPES.initial;
+  opts.requestType = REQUEST_TYPE.initial;
   return checkFinished(opts, [], '%token%');
 }
